Add KMQ shard status page
