+++ package.json
@@ -36,6 +36,7 @@
     "commander": "^7.0.0",
     "cron-parser": "^3.1.0",
     "dotenv": "^10.0.0",
+    "ejs": "^3.1.6",
     "eris": "https://github.com/Brainicism/eris.git#alpha-09-02-21",
     "eris-fleet": "https://github.com/danclay/eris-fleet#dev",
     "eris-pagination": "https://github.com/brainicism/eris-pagination#81aa2f0",
@@ -48,6 +49,7 @@
     "moment-timezone": "^0.5.32",
     "mysql2": "^2.2.5",
     "node-schedule": "^2.0.0",
+    "point-of-view": "^4.15.2",
     "sinon": "^11.1.1",
     "sodium-native": "^3.2.1",
     "ts-node": "^10.0.0",
