+++ src/helpers/management_utils.ts
@@ -1,7 +1,6 @@
 /* eslint-disable global-require */
 /* eslint-disable import/no-dynamic-require */
 import schedule from "node-schedule";
-import fastify from "fastify";
 import _ from "lodash";
 import { IPCLogger } from "../logger";
 import { state } from "../kmq";
@@ -27,7 +26,6 @@ import guildDeleteHandler from "../events/client/guildDelete";
 import unavailableGuildCreateHandler from "../events/client/unavailableGuildCreate";
 import guildAvailableHandler from "../events/client/guildAvailable";
 import interactionCreateHandler from "../events/client/interactionCreate";
-import { userVoted } from "./bot_listing_manager";
 import { chooseRandom } from "./utils";
 import { reloadFactCache } from "../fact_generator";
 import MessageContext from "../structures/message_context";
@@ -74,29 +72,6 @@ export function registerProcessEvents() {
     process.on("unhandledRejection", unhandledRejectionHandler)
         .on("uncaughtException", uncaughtExceptionHandler)
         .on("SIGINT", SIGINTHandler);
-}
-
-/** Starts web server */
-export async function startWebServer() {
-    const httpServer = fastify({});
-    httpServer.post("/voted", {}, async (request, reply) => {
-        const requestAuthorizationToken = request.headers["authorization"];
-        if (requestAuthorizationToken !== process.env.TOP_GG_WEBHOOK_AUTH) {
-            logger.warn("Webhook received with non-matching authorization token");
-            reply.code(401).send();
-            return;
-        }
-
-        const userID = request.body["user"];
-        await userVoted(userID);
-        reply.code(200).send();
-    });
-
-    try {
-        await httpServer.listen(process.env.WEB_SERVER_PORT, "0.0.0.0");
-    } catch (err) {
-        logger.error(`Erroring starting HTTP server: ${err}`);
-    }
 }
 
 /**
